EventPtr's are now created and managed using an EventFactory class. Currently, this just calls new/delete, but it's poised to soon perform more efficient memory management.
Added PionLockedQueue to the XCode project

+++ platform/tests/ConfigServiceTests.cpp
@@ -245,7 +245,7 @@ public:
 		BOOST_CHECK_EQUAL(response_ptr->getStatusCode(), HTTPTypes::RESPONSE_CODE_NO_CONTENT);
 
 		// make sure there's no content
-		BOOST_CHECK_EQUAL(response_ptr->getContentLength(), 0);
+		BOOST_CHECK_EQUAL(response_ptr->getContentLength(), static_cast<size_t>(0));
 	}
